LayoutTests/imported/w3c/web-platform-tests/storage/storagemanager-estimate.https.any-expected.txt


PASS estimate() method returns a Promise
PASS estimate() resolves to dictionary with members
PASS estimate() shows usage increase after 1MB IndexedDB record is stored
